requirements.txt
dash
dash_bootstrap_components
fastapi
uvicorn
gunicorn
httpx
openai
python-dotenv
pandas
jinja2
pydantic
google-cloud-firestore
PyYAML
